### 2024-02-26 #236 @paocorrales
Hi! It is good to have news form JOSE :tada:
The paper is up to date, nothing has changed since November. Looking forward to the review process!
