### 2024-03-03 #236 @yabellini
> For future reference, @yabellini, you can use the `assign` command like I did above!

Good to know. I saw the command but wasn’t clear if I could assign myself. Thanks!
