### 2024-08-23 #236 @yabellini
Hi, @dpshelio. You volunteer to be a reviewer for JOSE.  I would like to know if you can review this submission.  As teaching is one of your topics, I think you are a good candidate for this work.
