### 2024-08-23 #236 @yabellini
Hi, @kbodwin. You volunteer to be a reviewer for JOSE. I would like to know if you can review this submission. Because of your experience with R and teaching, I think you are a great candidate to review this work.
