### 2024-08-23 #236 @mdogucu
> Hi @mdogucu, you volunteered to be a reviewer for JOSE. We would like to know if you are willing to review this submission.

I am sorry but I am unable to review at this time. Could I also be removed from the reviewer list? I had signed up seven years ago but current life circumstances have changed. Thank you for understanding!
