### 2024-09-16 #260 @editorialbot
[REVIEW]: An R reproducibility toolkit for the practical researcher
**Submitting author:** @paocorrales (<a href="http://orcid.org/0000-0003-1923-9129">Paola Corrales</a>)
**Repository:** https://github.com/eliocamp/reproducibility-with-r/
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.3
