### 2024-09-18 #260 @Aariq
- The Turing Way is referenced with a link to a zenodo record rather than the DOI for that record.  Might be better to use https://doi.org/10.5281/zenodo.2669548

These are great materials and the short paper introduces them well.  I teach a workshop series that covers similar topics and I'm very excited to adopt some of the materials. I think the guide is written very thoughtfully with their target audience in mind.  I especially appreciated the discussion of what reproducibility is and how it is not necessarily related to correctness and isn't a binary. As I was going through the materials on reproducibility.rocks, I opened a few minor issues:

- https://github.com/eliocamp/reproducibility-with-r/issues/24
- https://github.com/eliocamp/reproducibility-with-r/issues/25
- https://github.com/eliocamp/reproducibility-with-r/issues/26
- https://github.com/eliocamp/reproducibility-with-r/issues/28

These issues mostly relate to content that is *slightly* out of date.  I'll also note that while it is true that Quarto is still rapidly evolving, I think very soon (if not already) it will be worth it to teach Quarto instead of RMarkdown for reproducible scientific manuscripts and reports.  RMarkdown and the `rticles` package certainly aren't going away anytime soon, but some of the frustrations I've encountered writing manuscripts in RMarkdown have been solved in Quarto—e.g. cross references to equations, images, code chunks, and other arbitrary sections—and I don't think Posit plans to actively develop RMarkdown going forwards. The authors may want to consider prioritizing an update to the [reproducible reporting section](https://reproducibility.rocks/materials/day1/03-rmarkdown/) in the next revision of these materials, although I do not think this update is necessary for acceptance of the manuscript.
