### 2024-09-18 eliocamp/reproducibility-with-r#24 @Aariq
Update RStudio download link
On Day 0 the link to https://www.rstudio.com/products/rstudio/download/ redirects to https://posit.co/products/, which is not exactly the right place to download RStudio.  I suggest updating to https://posit.co/download/rstudio-desktop/

(part of JOSE review here: https://github.com/openjournals/jose-reviews/issues/260)
