### 2024-09-18 eliocamp/reproducibility-with-r#25 @Aariq
next and previous page links don't go where I'd expect
At the bottom of the Day 0 page, the "next page" link goes to the mini reprohack page, not Day 1 like I would have expected.  The previous page on the mini reprohack page goes back to Day 0.

On the [main day 1 page](https://reproducibility.rocks/materials/day1/), there is no next or previous page link (this applies to the main page for day 2–4 also)

(part of JOSE review here: https://github.com/openjournals/jose-reviews/issues/260)
