### 2024-09-18 eliocamp/reproducibility-with-r#26 @Aariq
link to Quarto journal article templates
I appreciate the "What about Quarto?" callout box on [this page](https://reproducibility.rocks/materials/day1/03-rmarkdown/#r-markdown-files), but you might link specifically to https://quarto.org/docs/extensions/listing-journals.html where readers can find journal article extensions as new ones are added.  I don't know if everything offered by `rticles` is there, but I suspect it soon will be.

(part of JOSE review here: https://github.com/openjournals/jose-reviews/issues/260)
