### 2024-09-18 eliocamp/reproducibility-with-r#28 @Aariq
RStudio & markdown bug resolved—update text?
I noticed that the [bug ](https://github.com/rstudio/rstudio/issues/9861)you link to in [this section](https://reproducibility.rocks/materials/day3/01-renv/#using-renv-in-your-project) is now closed.  Probably not urgent, but next time you revise this text, you might remove this part.
