### 2024-09-22 b8f2226
Mentions that renv-rmarkdown bug was for previous RStudio version
Closes #28
