### 2024-09-22 eliocamp/reproducibility-with-r#28 @eliocamp
Thanks. I've confirmed that the bug not longer happens with the current RStudio version. I kept the message in case someone's using an old RStudio version.
