### 2024-09-22 106c9f9
Changes rstudio download location
Closes #24
