### 2024-09-22 ebad3b4
Links to quarto's journal templates
Closes #26
