### 2024-09-22 eliocamp/reproducibility-with-r#25 @eliocamp
Yeah, this is not ideal. I tried looking into this and it seems to be a limitation with the apero theme we're using since their example site has the same problems. I [opened an issue](https://github.com/hugo-apero/hugo-apero/issues/114) upstream to see if this is something that can be fixed site-side or if it should be fixed on their end.
