### 2024-09-22 #260 @eliocamp
Thanks for the review and the open issues. I've fixed 3/4. The missing one seems to be a limitation of the theme we're using for the website. I've open an issue in that repository to see what we can do.

And yes on quarto. We hadn't looked at the list of available journal templates for a while, so we didn't know it was so large (although not as large as rticles). We would need to look how to use a custom LaTeX template with quarto.
