### 2024-10-16 #260 @yabellini
Hello @luisDVA, thank you so much for agreeing to review this proposal 🙏 .

I added you as a reviewer. You can generate your checklist by calling our bot:

`@ editorialbot generate my checklist`
