### 2024-10-22 c7ddb23
Small update bios and preparation
