+++ content/people/paocorrales/index.md
@@ -35,7 +35,7 @@ links:
   url: https://www.instagram.com/paobcorrales
 ---
 
-I have a PhD from the University of Buenos Aires. I study atmospheric sciences applying data assimilation techniques to improve short-term forecasts of severe events in Argentina. Also, I'm a Trainer and instructor for The Carpentries and an RStudio certified instructor. I'm also part of Expedition Science, an Argentina-based NPO, where she leads educational projects such as science camps and workshops for students and K-12 science teachers. I'm a professor at the Data Sciences degree and postgraduate courses at Guillermo Brown University.
+I have a PhD from the University of Buenos Aires. I study atmospheric sciences applying data assimilation techniques to improve short-term forecasts of severe events in Argentina. Also, I'm a Trainer and instructor for The Carpentries and an RStudio certified instructor. I'm also part of Expedition Science, an Argentina-based NPO, where she leads educational projects such as science camps and workshops for students and K-12 science teachers. I'm a professor at the Data Sciences degree at Austral University.
 
 I develops openly licensed materials to teach and learn R from scratch.
 
