+++ content/materials/slides/day2/index.qmd
@@ -2,6 +2,8 @@
 title: "Day 2"
 format: 
   revealjs:
+    theme:
+      - ../custom_slides.scss
     logo: "logo.png"
     title-slide-attributes:
       data-background-image: "https://reproducibility.rocks/sheep.svg"
