### 2024-10-27 #260 @paocorrales
Hello! We'll be teaching the workshop this week and are making the usual checks and updates to the material. We haven't changed anything big, just broken links and small additions.
I wanted to mention it here in case anyone is currently visiting the web or the repo!
