+++ content/materials/day0/index.Rmd
@@ -155,9 +155,6 @@ The core steps are:
 On the fourth day we show the basics of docker.
 Instructions on how to install it depend on your operating system.
 
-On the fourth day we show the basics of docker.
-Instructions on how to install it depend on your operating system.
-
 If you're on Windows, you need to enable hardware virtualisation. 
 This needs to be enabled on the BIOS ([how to do this varies between manufacturers](https://bce.berkeley.edu/enabling-virtualization-in-your-pc-bios.html)) and as a Windows' setting. 
 Search for "Turn Windows Features on and off" and make sure that "Virtual Machine Platform",  "Windows Hypervisor Platform" and "Windows Subsystem for Linux" are enabled. 
