### 2024-11-01 #260 @luisDVA
### Review
I enjoyed reviewing the contents and materials for this teaching module. The major topics are all highly relevant to reproducibility in scientific research, and the content for each day is laid out with a suitable pace and easy-to-follow exercises and examples. The course motivation and outline are also very well summarized in the corresponding short document.

I pointed out various minor fixes and made suggestions in the following issues:
- [#31](https://github.com/eliocamp/reproducibility-with-r/issues/31)
- [#32](https://github.com/eliocamp/reproducibility-with-r/issues/32)
- [#33](https://github.com/eliocamp/reproducibility-with-r/issues/33)
- [#34](https://github.com/eliocamp/reproducibility-with-r/issues/34)
- [#35](https://github.com/eliocamp/reproducibility-with-r/issues/35)
- [#36](https://github.com/eliocamp/reproducibility-with-r/issues/36)

Overall, the teaching module is cohesive, relevant, and with great potential for reuse. The other reviewer above already noted that some course elements that revolve around RMarkdown are at potential risk of becoming outdated or somewhat overlooked in favor of the Quarto publishing framework. This is already being addressed, but worth noting again.

The section on reporting with RMarkdown (day 1, part 3) is very well planned in terms of how additional topics and tools are introduced without overwhelming the learners with complexity. Since these materials have now been used to teach the course in repeated occasions, I wonder if there have been issues with LaTeX templates, and if so, how these were solved. If this is the case, perhaps this could be added to the instructor notes.

While I appreciate the power of using research compendia (day 2, part 3), this practice is not really widespread (at least within my field of biology/ecology/evolution). To get the most out of this tooling and approach, I suggest adding more explicit examples to the materials, and to dedicate some time to look at research compendia produced for published journal articles.

The materials for the fourth day are particularly valuable, given the clear and easy to follow instructions for setting up docker containers and even mapping folders from the virtual environments to the local machine. This content is rarely taught in research settings in biological science (the general target for the course), and the focus on R and running the IDE in the container is quite useful.

I appreciate the invitation to review and hope that the module authors find my comments useful.
