### 2024-11-25 eliocamp/reproducibility-with-r#31 @luisDVA
day 0: consider `usethis::git_sitrep`
When setting up all the necessary components for using git with R, consider suggesting that the learners run `git_sitrep` once the setup is complete, to get a nice overview of if/how git is ready to go

this is part of my [review](https://github.com/openjournals/jose-reviews/issues/260) for JOSE
