### 2024-11-25 eliocamp/reproducibility-with-r#32 @luisDVA
day 1, topic 2. minor edits
the shortcut in the callout box for restarting the R session says "shif" instead of "shift", and this reminded to suggest mentioning the differences in keyboard shortcuts (ctrl/command keys, alt/option keys)  to cover windows and mac users. Either here or somewhere relevant.

below this the text says "be default RStudio" instead of "by default RStudio"

- these are just minor typing fixes but the windows/mac note may be more relevant

part of my JOSE [review](https://github.com/openjournals/jose-reviews/issues/260)
