### 2024-11-25 eliocamp/reproducibility-with-r#33 @luisDVA
day 2, topic 2: git - broken link
Great section with lots of very nice diagrams in terms of explaining concepts and also in their styling and design.
Heads up - The link to the workshop by Yani and Marysol is broken

---
part of my [JOSE review](https://github.com/openjournals/jose-reviews/issues/260)
