### 2024-11-28 eliocamp/reproducibility-with-r#34 @luisDVA
day 1, 03. RMarkdown. Minor edits and figure suggestion.
In some places _header_ (in reference to the yaml preamble) is spelled _"heather"_. Just a heads up.

When introducing templates for different journals, consider adding a composite figure to the materials showing the same input document rendered for various journals, as a more visual representation of the benefit of reproducible reporting using templates.

part of my [JOSE review](https://github.com/openjournals/jose-reviews/issues/260)
