### 2024-11-28 eliocamp/reproducibility-with-r#35 @luisDVA
day 4. update links for web-based platforms
[JOSE review](https://github.com/openjournals/jose-reviews/issues/260)

When mentioning web platforms in the materials, consider updating the links.
First, even though the redirect is working, replace RStudio Cloud with Posit Cloud to avoid confusion. Also, Digital Ocean may be worth mentioning here as it has some useful images for working with RStudio online as well as the analogsea R helper package.
