### 2024-11-28 eliocamp/reproducibility-with-r#36 @luisDVA
day 03, 02. data sharing. typos and licensing suggestions
This is a very relevant topic and I really liked the writing for this section.

I noticed some typos like "lis" in the caption for fig 5, and multiple misspellings of Zenodo as Zeonodo.

When discussing data sharing, in the context of these materials it may we worth adding some sentences about licenses for datasets, possibly mentioning the [Open Data Commons](https://opendatacommons.org/licenses/by/) license, championed by the Open Knowledge Foundation and potentially relevant here.

part of my [JOSE review](https://github.com/openjournals/jose-reviews/issues/260]
