### 2024-11-29 140d95f
Add git_sitrep() Closes #31
