### 2024-11-29 f808905
Fix typo and add shorcuts for mac
Closes #32
