### 2025-01-03 c6f4f5e
Adds rticles comparison
#34
