+++ content/materials/slides/day3/index.qmd
@@ -13,7 +13,7 @@ format:
 ---
 
 # Day 3:
-
+.
 1.  [Managing R dependencies with renv](/materials/day3/01-renv/)
 2.  [Sharing your data](/materials/day3/02-data/)
 
