### 2025-01-03 eliocamp/reproducibility-with-r#34 @eliocamp
> In some places _header_ (in reference to the yaml preamble) is spelled _"heather"_. Just a heads up.

This was probably in honour of  Heather Turner :P.
