### 2025-01-03 c343c06
Adds discussion of data licenses.
#36
