### 2025-01-03 6cec249
Changes description of licence page
