### 2025-01-03 #260 @eliocamp
Than you, @luisDVA .

We added an example of a research compendium in the materials and a few lines about it.

That section is tricky because we don't have enough time to go into how to build an R package so we do a very superficial explanation. Novice students get a bit intimidated by the prospect and advanced students don't get a lot out of it.  For the latest version of the course, we added a few [slides](https://reproducibility.rocks/materials/slides/day2/index.html#/rrtools-level-1-project-structutre) trying to explain the various levels explicitly.

We will probably rework this section if we get to teach the course again.

We've closed all the outstanding issues with fixes. Let me know if there's anything else.
