### 2025-01-03 #260 @yabellini
Thanks @eliocamp and @paocorrales for your work on the reviewers sugestiones.

@Aariq and @luisDVA please let me know if you are ok with the changes made and if I can recommend the acceptance of the paper.

Thank you all!
