### 2025-01-06 #260 @luisDVA
Hi @yabellini, all suggestions were addressed satisfactorily and all the updated content looks great.
