### 2025-01-07 #260 @yabellini
Please @paocorrales , @eliocamp update the versión of the course by making a New release with all the changes and make a Zenodo archive. Please report the DOI of that new versión and the versión here.
