### 2025-01-07 #260 @yabellini
Also, please @paocorrales and @eliocamp check for typos and if you need to update some other information on the paper (like affiliations).
