+++ jose-submission/paper.md
@@ -34,7 +34,7 @@ The workshop is available at [reproducibility.rocks](http://reproducibility.rock
 
 # Statement of Need
 
-Reproducibility is should be part of any data analysis work, but it's particularly important in scientific research.
+Reproducibility should be part of any data analysis work, but it's particularly important in scientific research.
 Achieving full reproducibility is challenging and requires, time, resources and specialised tools and skills [@whitaker2019].
 This workshops empowers researchers so they can maximize the reproducibility of their work by leveraging existing tools in the R ecosystem --such as RMarkdown, renv, and others--, version control and containers.
 
@@ -46,16 +46,16 @@ The [notes for instructors](https://reproducibility.rocks/notes_for_instructors/
 The workshop is intended for researchers, data scientists, and anyone who uses R to generate documents and who wants to collaborate with other people (or themselves in the future).
 Basic prior experience with R is highly recommended.
 
-The materials are organized in four 3-hour modules, ideally covered over a few weeks to give time to spaced practice [@benjamin2010] and homework assessment.
+The materials are organized in four 3-hour modules, ideally covered over a few weeks to allow time for spaced practice [@benjamin2010] and homework assessment.
 Each module has specific learning objectives (Table 1) and associated homework with the common goal of taking an existing non-reproducible research project and pushing it further in the reproducibly spectrum with each step.
 
+| Day \# | Learning Objectives | Topics |
+|------------------|-------------------------------------|------------------|
+| [Day 1](https://reproducibility.rocks/materials/day1/): The workbench | 1\. Create an R project with a specific and useful structure; 2. Extend the use of RMarkdown files for reports and papers; 3. Use LaTeX templates to customize RMarkdown reports. | 1\. What is reproducibility anyway?; 2. A tidy project is a happy project 3. Reporting with R Markdown |
+| [Day 2](https://reproducibility.rocks/materials/day2/): Sharing and Contributing | 1\. Use git to track your work; Use git to collaborate with others; How to set up project documentation; Organizing a project. | 1\. here are the portable paths, 2. Getting the most of git; 3. Research compendia with rrtools |
+| [Day 3](https://reproducibility.rocks/materials/day3/): Managing Dependencies and data | 1\. Manage package dependencies inside a project; 2. How to pack and share data associated wth a project. | 1\. Managing R dependencies with renv; 2. Sharing data |
+| [Day 4](https://reproducibility.rocks/materials/day4/): Building Reproducible Computational Environments with Docker | 1\. Using docker to share a project. | 1\. Reproducible computational environment with Docker |
 
-| Day \#                                                                                                               | Learning Objectives                                                                                                                                                            | Topics                                                                                                 |
-|-----------------|--------------------------------------|-----------------|
-| [Day 1](https://reproducibility.rocks/materials/day1/): The workbench                                                | 1\. Create R project with an specific and useful structure; 2. Extend the use of RMarkdown files for reports and papers; 3. Use LaTeX templates to customize RMarkdown reports | 1\. What is reproducibility anyway?; 2. A tidy project is a happy project 3. Reporting with R Markdown |
-| [Day 2](https://reproducibility.rocks/materials/day2/): Sharing and Contributing                                     | 1\. Using git to track your work; Using git to collaborate with others; How to set the documentation of your project; Organizing a project                                     | 1\. here are the portable paths, 2. Getting the most of git; 3. Research compendia with rrtools        |
-| [Day 3](https://reproducibility.rocks/materials/day3/): Managing Dependencies and data                               | 1\. Manage package dependencies inside a project; 2. How to pack and share data associated to a project                                                                                    | 1\. Managing R dependencies with renv; 2. Sharing data                                                 |
[…]
