### 2025-01-13 #260 @eliocamp
Thanks!
We've corrected a few typos and grammar errors and created a new release. This is the zenodo record: https://zenodo.org/records/14636061
