### 2025-01-13 #260 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
