### 2025-01-13 #260 @paocorrales
Hi Yani!

I'm not sure if I understand this summary correctly, just in case:
* Reference 1 already include the doi
* Reference 2 only has the link to a zenodo record but we could add the doi https://doi.org/10.5281/zenodo.2669548

> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> ✅ OK DOIs
>
> - 10.1016/j.cogpsych.2010.05.004 is OK
>
> 🟡 SKIP DOIs
>
> - No DOI given, and none found for title: The Turing Way: Sharing the Responsibility of Repr...
> - No DOI given, and none found for title: Spaced Practice
>
> ❌ MISSING DOIs
>
> - None
>
> ❌ INVALID DOIs
>
> - None
> ```
