### 2025-01-13 #260 @yabellini
@paocorrales I see the doi in the first Benjamin et al. reference, and it would be good to add the DOI to The Turing Way reference.  Please make those changes and let me know so I can generate the new document.
