### 2025-01-13 #260 @yabellini
The Zenodo link also works, so If you want to leave it as is, it is also ok.
