### 2025-01-25 #260 @eliocamp
I've added the DOI, should I create a new release?
