### 2025-02-03 #260 @yabellini
Hi @eliocamp @paocorrales do you have news about the release? It would be great so we can publish and close this issue :-)
