### 2025-02-05 #260 @eliocamp
[v2.0.2](https://github.com/eliocamp/reproducibility-with-r/releases/tag/v2.0.2)
