### 2025-02-05 #260 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
