### 2025-02-05 #260 @yabellini
The Editor in Chief will handle the last step before the publication. Congrats! 😄
