### 2025-04-16 #260 @labarba
We request that authors edit the metadata of the Zenodo deposit so **title** and author list match the JOSE paper. Could you do that, please?
