### 2025-04-16 #260 @labarba
Comments:
- p.1, you cite a slide deck after a general statement about reproducibility taking time and effort; it seems random; could you perhaps find a more appropriate reference, or remove the citation altogether since it's a well-known fact?
- "This workshops empowers…" >> workshop
- "The notes for instructors has…" >>have
- p.2, "we encourage self-thought students" >>self-taught
- p.3, "requests from the students were include" >>included

(Such a pretty website!!)
