### 2025-04-18 #260 @labarba
I don't see that the Zenodo archive **title** has been updated, as requested [above](https://github.com/openjournals/jose-reviews/issues/260#issuecomment-2810685979)?
